unidad-2 | update vagrant boostrap

+++ Vagrant.bootstrap.sh
@@ -85,28 +85,44 @@ pwd
 echo " "
 ls -ltra
 
-sudo docker ps -a | grep -v CONTAINER | while read line ; do sudo docker rm `echo $line | awk '{print$1}'` ; done
-sudo docker images | grep -v REPOSITORY | while read line ; do sudo docker image rm  `echo $line | awk '{print$3}'` ; done
+echo " ========================================================================" 
+echo " "
+echo " reconstruyendo docker"
+if [ `sudo docker ps | wc -l` -gt 1 ] ; then  
+	sudo docker ps | grep -v CONTAINER | awk '{print$1}' | while read line ; do sudo docker stop $line ; done
+	sudo docker ps -a | grep -v CONTAINER | while read line ; do sudo docker rm `echo $line | awk '{print$1}'` ; done
+	sudo docker images | grep -v REPOSITORY | while read line ; do sudo docker image rm  `echo $line | awk '{print$3}'` ; done
+fi
 
 echo " "
 echo "creando docker"
 sudo docker-compose up -d
 
 echo " "
-echo "docker creados"
+echo "docker activos "
 sudo docker ps
+echo "docker creados "
 sudo docker ps -a
 
-echo " ------------------------------------------------------------- "
+echo " ========================================================================"
 echo "IP DE BASE DE DATOS"
 sudo docker inspect `sudo docker ps | grep mysql | awk '{print$1}'` | grep IPAddress | tail -1 | awk '{print$2}' | sed 's/\"//g' | sed 's/\,//g'
-
-echo " ------------------------------------------------------------- "
+IP_BD=`sudo docker inspect \`sudo docker ps | grep mysql | awk '{print$1}'\` | grep IPAddress | tail -1 | awk '{print$2}' | sed 's/\"//g' | sed 's/\,//g'`
+echo " ========================================================================"
 echo "IP WEB SERVER"
 sudo docker inspect `sudo docker ps | grep php | awk '{print$1}'` | grep IPAddress | tail -1 | awk '{print$2}' | sed 's/\"//g' | sed 's/\,//g'
 
-
+echo " ========================================================================"
+echo "Adecuando db_connect a la IP del contenedor de BD"
+sudo sed -i 's/127.0.0.1/'${IP_DB}'/' /var/www/myapp/src/include/db_connect.php
 #echo " "
 #sudo docker exec -i apache2_php cd /var/www/html/myapp
 
-sudo docker exec -i db_mysql mysql -uroot -proot devops_app < /vagrant/docker/configs/mysql/script.sql
+echo " ========================================================================"
+
+echo "creando DB"
+if [ `sudo su - ; ls /var/db/mysql/devops_app | grep welcome | wc -l` -gt 2  ] ; then
+	echo "db y tabla creada"
+else
+	sudo docker exec -i db_mysql mysql -uroot -proot devops_app < /vagrant/docker/configs/mysql/script.sql
+fi
